# Small performance improvements