The codec interface has been modified to allow the reuse of buffers when decoding
documents one by one from a table.

Memory consumption has been reduced drastically (-60%) and iteration speed has been improved by 20%

```
name                          old time/op    new time/op    delta
PreparedSelectWhere/00001-12    8.28µs ± 0%    7.24µs ± 0%  -12.51%  (p=1.000 n=1+1)
PreparedSelectWhere/00010-12    14.3µs ± 0%    12.7µs ± 0%  -11.16%  (p=1.000 n=1+1)
PreparedSelectWhere/00100-12    91.7µs ± 0%    77.3µs ± 0%  -15.65%  (p=1.000 n=1+1)
PreparedSelectWhere/01000-12     862µs ± 0%     728µs ± 0%  -15.55%  (p=1.000 n=1+1)
PreparedSelectWhere/10000-12    9.08ms ± 0%    7.34ms ± 0%  -19.13%  (p=1.000 n=1+1)

name                          old alloc/op   new alloc/op   delta
PreparedSelectWhere/00001-12    1.64kB ± 0%    1.69kB ± 0%   +3.43%  (p=1.000 n=1+1)
PreparedSelectWhere/00010-12    2.93kB ± 0%    1.83kB ± 0%  -37.47%  (p=1.000 n=1+1)
PreparedSelectWhere/00100-12    15.9kB ± 0%     3.3kB ± 0%  -79.41%  (p=1.000 n=1+1)
PreparedSelectWhere/01000-12     146kB ± 0%      18kB ± 0%  -87.85%  (p=1.000 n=1+1)
PreparedSelectWhere/10000-12    1.45MB ± 0%    0.16MB ± 0%  -88.76%  (p=1.000 n=1+1)

name                          old allocs/op  new allocs/op  delta
PreparedSelectWhere/00001-12      32.0 ± 0%      30.0 ± 0%   -6.25%  (p=1.000 n=1+1)
PreparedSelectWhere/00010-12      86.0 ± 0%      48.0 ± 0%  -44.19%  (p=1.000 n=1+1)
PreparedSelectWhere/00100-12       626 ± 0%       228 ± 0%  -63.58%  (p=1.000 n=1+1)
PreparedSelectWhere/01000-12     6.03k ± 0%     2.03k ± 0%  -66.35%  (p=1.000 n=1+1)
PreparedSelectWhere/10000-12     60.0k ± 0%     20.0k ± 0%  -66.64%  (p=1.000 n=1+1)
```